Translation to spanish
Hi!

This is the first time collaborating with CakePHP. I want to translate some docs to Spanish, in order to help adoption of the framework for Spanish-speakers developers. Please note the:

&aacute; for á
&eacute; for é
&iacute; for í
&oacute; for ó
&uacute; for ú

Html entities. I assumed this is the best way to represent this non-standard characters, but if there is a better way, please let me know and I'll correct the proposal.

Saludos, (greetings)